+++ spark/src/test/scala/org/apache/comet/parquet/CometParquetWriterSuite.scala
@@ -949,7 +949,7 @@ class CometParquetWriterSuite extends CometTestBase {
       var iterations = 0
 
       while (capturedPlan.isEmpty && iterations < maxIterations) {
-        Thread.sleep(checkIntervalMs)
+        Thread.sleep(checkIntervalMs.toLong)
         iterations += 1
       }
